# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API GameDashboard",
    "version": "1.0.0",
    "description": "Panel de gestion de serveurs de jeu. Les erreurs suivent Problem Details (RFC 9457) : un corps JSON portant `title`, `status` et `detail`, et non un simple texte."
  },
  "servers": [
    {
      "url": "https://panel.example",
      "description": "Le panel, sur son domaine"
    }
  ],
...
  "paths": {
...
    "/api/v1/client/servers": {
      "get": {
        "summary": "Lister les serveurs auxquels la clé donne accès.",
        "description": "Clé personnelle.",
        "tags": [
          "Serveurs"
        ],
        "operationId": "client-get-servers",
        "security": [
          {
            "clientKey": []
          }
        ],
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans le droit sur cette ressource."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
    "/api/v1/client/servers/{server}": {
      "get": {
        "summary": "Détail d'un serveur : limites, allocation, egg, état.",
        "description": "Clé personnelle.",
        "tags": [
          "Serveurs"
        ],
        "operationId": "client-get-servers-by-server",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "description": "Identifiant du serveur, ou son identifiant court.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "clientKey": []
          }
        ],
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans le droit sur cette ressource."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
...
    "/api/v1/application/domains/certificates": {
      "get": {
        "summary": "Domaines vérifiés et état de leur certificat TLS. « pending » dit s'il y a quelque chose à faire : aucun certificat, expiration à moins de trente jours, ou échec vieux d'une heure.",
        "description": "Clé de plateforme. Portée requise : `domains.certificates`.",
        "tags": [
          "Certificats"
        ],
        "operationId": "application-get-domains-certificates",
        "security": [
          {
            "applicationKey": [
              "domains.certificates"
            ]
          }
        ],
        "x-gd-scope": "domains.certificates",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « domains.certificates »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
    "/api/v1/application/domains/{domain}/certificate": {
      "post": {
        "summary": "Rendre compte d'une tentative. « issuedAt » et « expiresAt » à la réussite, « failure » à l'échec — et les deux ensemble quand un renouvellement échoue alors que l'ancien certificat tient encore.",
        "description": "Clé de plateforme. Portée requise : `domains.certificates`.",
        "tags": [
          "Certificats"
        ],
        "operationId": "application-post-domains-by-domain-certificate",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "description": "Valeur de « domain ».",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "domains.certificates"
            ]
          }
        ],
        "x-gd-scope": "domains.certificates",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « domains.certificates »."
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      }
    },
    "/api/v1/application/identity": {
      "get": {
        "summary": "Vérifier la clé, l'adresse IP et les portées accordées. Ne crée rien.",
        "description": "Clé de plateforme.",
        "tags": [
          "Mise en service"
        ],
        "operationId": "application-get-identity",
        "security": [
          {
            "applicationKey": []
          }
        ],
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans le droit sur cette ressource."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
    "/api/v1/application/users": {
      "get": {
        "summary": "Retrouver un compte par identifiant externe, adresse e-mail ou identifiant local.",
        "description": "Clé de plateforme. Portée requise : `users.read`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-get-users",
        "parameters": [
          {
            "name": "externalId",
            "in": "query",
            "required": false,
            "description": "Valeur de « externalId ».",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.read"
            ]
          }
        ],
        "x-gd-scope": "users.read",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.read »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "post": {
        "summary": "Créer un compte client. Aucun mot de passe n'est accepté : la connexion passe par le SSO.",
        "description": "Clé de plateforme. Portée requise : `users.write`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-post-users",
        "security": [
          {
            "applicationKey": [
              "users.write"
            ]
          }
        ],
        "x-gd-scope": "users.write",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.write »."
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      }
    },
    "/api/v1/application/users/{user}": {
      "get": {
        "summary": "Détail d'un compte et nombre de serveurs possédés.",
        "description": "Clé de plateforme. Portée requise : `users.read`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-get-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.read"
            ]
          }
        ],
        "x-gd-scope": "users.read",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.read »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "patch": {
        "summary": "Corriger nom, prénom ou identifiant externe. Ni rôle, ni adresse, ni mot de passe.",
        "description": "Clé de plateforme. Portée requise : `users.write`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-patch-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.write"
            ]
          }
        ],
        "x-gd-scope": "users.write",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.write »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      },
      "delete": {
        "summary": "Supprimer un compte. Refusé tant qu'il possède des serveurs.",
        "description": "Clé de plateforme. Portée requise : `users.delete`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-delete-users-by-user",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant du compte.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "applicationKey": [
              "users.delete"
            ]
          }
        ],
        "x-gd-scope": "users.delete",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.delete »."
          },
          "2XX": {
            "description": "Résultat de la lecture."
          }
        }
      }
    },
    "/api/v1/application/users/sso-link": {
      "post": {
        "summary": "Lien de connexion d'un client, désigné par userId ou externalId. Vaut deux minutes, sert une fois ; refusé pour le personnel.",
        "description": "Clé de plateforme. Portée requise : `users.sso`.",
        "tags": [
          "Comptes"
        ],
        "operationId": "application-post-users-sso-link",
        "security": [
          {
            "applicationKey": [
              "users.sso"
            ]
          }
        ],
        "x-gd-scope": "users.sso",
        "responses": {
          "400": {
            "description": "Demande refusée : la cause est dans le corps.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Aucune authentification recevable."
          },
          "403": {
            "description": "Authentifié, mais sans la portée « users.sso »."
          },
          "2XX": {
            "description": "Effectué."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-gd_session",
        "description": "Cookie de session du panel. Posé à la connexion, illisible par un script, et donc inutilisable depuis un client tiers."
      },
      "clientKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé personnelle, créée depuis « Compte → Clés d'API ». Elle agit au nom de son propriétaire et ne porte jamais plus que les portées qu'on lui a données."
      },
      "applicationKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé de plateforme, créée depuis l'administration. Elle n'agit au nom de personne : elle provisionne, suspend et résilie pour le compte d'un système tiers."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "Refus ou erreur, au format Problem Details (RFC 9457).",
        "properties": {
          "title": {
            "type": "string",
            "description": "Ce qui s'est passé, en une phrase."
          },
          "status": {
            "type": "integer",
            "description": "Le code HTTP, répété dans le corps."
          },
          "detail": {
            "type": "string",
            "description": "Le détail utile, quand il y en a un."
          }
        },
        "required": [
          "title",
          "status"
        ]
      }
    }
  }
}